metrics: make prometheus optional
